validation: ipsec: move combined out+in test code away from the infra
Move ipsec_check_out_in_one() from the generic framework file to a
test case file where it better belongs.
